% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_catchnum}
\alias{check_catchnum}
\title{Check if CATCHNUM column exist in catchments_sf table.}
\usage{
check_catchnum(catchments_sf)
}
\arguments{
\item{catchments_sf}{A single column name in input_table, or a vector of multiple column names.}
}
\value{
TRUE/FALSE.
}
\description{
Check if CATCHNUM column exist in catchments_sf table.
}
