% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_for_geometry}
\alias{check_for_geometry}
\title{Check if a sf objet as a geometry column.}
\usage{
check_for_geometry(in_sf)
}
\arguments{
\item{in_sf}{A sf object.}
}
\value{
TRUE/FALSE.
}
\description{
Check if a sf objet as a geometry column.
}
