% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{dissolve_catchments_from_table}
\alias{dissolve_catchments_from_table}
\title{Get lists of catchments from a table and dissolve.}
\usage{
dissolve_catchments_from_table(
  catchments_sf,
  input_table,
  out_feature_id,
  calc_area = FALSE,
  intactness_id = NULL,
  dissolve_list = c(),
  drop_table = NULL
)
}
\arguments{
\item{catchments_sf}{sf object of the catchments dataset with unique identifier column: CATCHNUM.}

\item{input_table}{Data frame where column names are polygon names and rows are catchments making up each polygon.
e.g. output from \code{\link[=get_upstream_catchments]{get_upstream_catchments()}}.}

\item{out_feature_id}{String representing the output column name holding the polygon unique identifiers.}

\item{calc_area}{If TRUE, an area_km2 column is added to the output containing the polygon area.}

\item{intactness_id}{Optional string identifying an intactness column (values between 0 and 1) in catchments_sf. If provided,
used to calculate the area weighted intactness (AWI) of the dissolved polygons.}

\item{dissolve_list}{Vector of columns in \code{input_table} to include in the output. Defaults to colnames(input_table).
Can also dissolve multiple columns from \code{input_table} together by combining column names with \code{__} (e.g. PB_0001__PB_0002).}

\item{drop_table}{A table in the same format \code{input_table} listing catchments to be dropped from the dissolve area.}
}
\value{
A sf object of polygons with unique identifier column \code{out_feature_id}.
}
\description{
Takes lists of catchments from a table where each column represents a polygon to be created, and dissolves the
catchments into a single polygon feature. Optionally adds area in km2, and area-weighted intactness.
}
\details{
\code{input_table} is in the form exported from the beaconsbuilder package, where each column lists catchment ids,
and each column name represents the unique identifier of the output polygon.

This function is used to create conservation area polygons using tables created by the beaconsbuilder package,
and upstream polygons using tables created by \code{\link[=get_upstream_catchments]{get_upstream_catchments()}}.

The optional \code{dissolve_list} parameter can be used to filter the column names in \code{input_table} that will
be dissolved. The \code{dissolve_list} can also contain combinations of column name separated by \code{__}, in this
case the output polygon will dissolve the combined area of catchments from all columns in the string. For example,
the string "PA1__PA2" in the \code{dissolve_list} will output a polygon representing the dissolved area of all
catchments listed in the PA1 and PA2 columns of \code{input_table}.

The main application of \code{drop_table} is to remove conservation area catchments from the dissolved area when dissolving
upstream catchments tables. See vignette('overview') for an example.
}
\examples{
library(sf)

# Individual conservation areas
conservation_areas <- dissolve_catchments_from_table(
  catchments_sf = catchments_sample,
  input_table = builder_table_sample,
  out_feature_id = "network",
  dissolve_list = c("PB_0001", "PB_0002"))
plot(conservation_areas)

# conservation areas combined into a network
network <- dissolve_catchments_from_table(
  catchments_sf = catchments_sample,
  input_table = builder_table_sample,
  out_feature_id = "network",
  dissolve_list = c("PB_0001__PB_0002"))
plot(network)

# Upstream polygons
upstream_table <- get_upstream_catchments(conservation_areas, "network", catchments_sample)
dissolve_catchments_from_table(
  catchments_sample,
  upstream_table,
  "network",
  TRUE)

}
