% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{extract_catchments_from_table}
\alias{extract_catchments_from_table}
\title{Get list of catchments from a table and extract to a new feature.}
\usage{
extract_catchments_from_table(
  catchments_sf,
  input_table,
  extract_feature_id,
  out_feature_id
)
}
\arguments{
\item{catchments_sf}{sf object of the catchments dataset with unique identifier column: CATCHNUM.}

\item{input_table}{Data frame where column names are polygon names and rows are catchments making up each polygon.
e.g. output from \code{\link[=get_upstream_catchments]{get_upstream_catchments()}}.}

\item{extract_feature_id}{String representing the column name to extrat from \code{input_table}. Can also take
a vector of names in which case multiple columns will be extracted.}

\item{out_feature_id}{String representing the output column name holding the polygon unique identifier.}
}
\value{
A sf object of catchments with unique identifier column \code{out_feature_id}.
}
\description{
Takes a list of catchments from a table column using the column name, and extracts the
catchments from the catchments dataset. If multiple column names are provided, the unique combination of
catchments from all columns is returned.
}
\examples{

extract_catchments_from_table(catchments_sample, builder_table_sample, "PB_0001", "network")
extract_catchments_from_table(
  catchments_sample, builder_table_sample, c("PB_0001", "PB_0002"), "network")
}
