% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_catch_list}
\alias{get_catch_list}
\title{Get unique sets of catchment values from a table.}
\usage{
get_catch_list(pa_ids, input_table)
}
\arguments{
\item{pa_ids}{A single column name in input_table, or a vector of multiple column names.}

\item{input_table}{Data frame where column names are protected area names and rows are catchments making
up each polygon. e.g. output from \code{\link[=get_upstream_catchments]{get_upstream_catchments()}}.}
}
\value{
A vector of unique catchment values.
}
\description{
Designed for use with wide format tables such as those output by \code{get_upstream_catchments} where each
column name is a unique protected area id, and each row is a catchment id. Catchment values from all columns
included in pa_ids are combined and returned after removing duplicates.
}
\examples{
get_catch_list("PB_0001", builder_table_sample)
get_catch_list(c("PB_0001", "PB_0003"), builder_table_sample)
}
