% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{long_to_wide}
\alias{long_to_wide}
\title{Convert long table into wide table with NAs (i.e. BUILDER style table)}
\usage{
long_to_wide(long_df, col_names, values_col)
}
\arguments{
\item{long_df}{A long style table with values in one column and group names in another.}

\item{col_names}{The column in long_df from which to get new column names.}

\item{values_col}{The column in long_df from which to get row values.}
}
\value{
A wide tibble
}
\description{
Used to prepare long tables for use in BUILDER or functions that take wide tables as
input (e.g.\code{dissolve_catchments_from_table()}).
}
\examples{
long_to_wide(
  long_df = wide_to_long(builder_table_sample, "group"),
  col_names = "group",
  values_col = "value")
}
