% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/names.R
\name{sep_network_names}
\alias{sep_network_names}
\title{Split network names into their constituent parts.}
\usage{
sep_network_names(network_names)
}
\arguments{
\item{network_names}{String of a network name, or a vector of multiple network names. Must include the \code{"__"} separator.}
}
\value{
A vector of names, or a list of vectors.
}
\description{
Takes a vector of network names in the form output by \code{\link[=gen_network_names]{gen_network_names()}} (i.e. \code{"__"} separated),
and splits the network names back into their original names.
}
\details{
If a single name is provided, a vector of names is returned. If a vector with length >1 is provided, a list
of vectors is returned with each list element named with the network name.
}
\examples{
sep_network_names("PB_1__PB_2")
sep_network_names(c("PB_1__PB_2", "PB_1__PB_3"))

x <- gen_network_names(c("PB_1", "PB_2", "PB_3", "PB_11"), 2)
sep_network_names(x)

x <- gen_network_names(c("PB_1", "PB_2", "PB_3", "PB_11"), 4)
sep_network_names(x)
}
